feat(controller): refactor CustomerController to use CustomerService and update endpoint paths feat(model): add Endereco class as embeddable for customer address details
