--- backend/customer/src/main/java/com/ms/customer/service/CustumerService.java
@@ -1,8 +1,0 @@
-package com.ms.customer.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class CustumerService {
-    
-}
